backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
motor==3.3.2
pymongo==4.6.1
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic[email]==2.4.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.1
numpy==1.26.0
python-dateutil==2.8.2
email-validator==2.1.0
